requirements.txt
fastapi==0.128.0
uvicorn==0.40.0

openai==2.14.0

pydantic==2.12.5
pydantic_core==2.41.5
typing_extensions==4.15.0
typing-inspection==0.4.2
annotated-types==0.7.0

python-dotenv==1.0.1

httpx==0.28.1
httpcore==1.0.9
anyio==4.12.0
sniffio==1.3.1
h11==0.16.0
idna==3.11
certifi==2025.11.12
distro==1.9.0
tqdm==4.67.1
click==8.3.1
starlette==0.50.0
jiter==0.12.0
